Consolidate library page handler into a single function.
