--- kolibri/plugins/learn/assets/src/modules/topicsRoot/index.js
@@ -1,14 +1,0 @@
-export default {
-  namespaced: true,
-  state: {
-    rootNodes: [],
-  },
-  mutations: {
-    SET_STATE(state, payload) {
-      state.rootNodes = payload.rootNodes;
-    },
-    RESET_STATE(state) {
-      state.rootNodes = [];
-    },
-  },
-};
